feature: complete set device data
